Translating Numbers To Java
Given a number, please translate it to a string, following the rules: 1 is translated to 'a', 2 to 'b', …, 12 to 'l', …, 26 to 'z'. For example, the number 12258 can be translated to 'aveh', 'aveh', 'abyh', 'lbeh' and 'lyh', so there are 5 different ways to translate 12258. How to write a function/method to count the different ways to translate a number in java